Finally time to eval. Cleaned up header file

+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: ohearn <ohearn@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/26 11:22:39 by ohearn        #+#    #+#                 */
-/*   Updated: 2022/12/17 20:03:34 by ohearn        ########   odam.nl         */
+/*   Updated: 2023/02/09 16:51:54 by ohearn        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 #include <unistd.h>
 #include <stdlib.h>
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	result;
-	int	negative;
+	long int	result;
+	int			negative;
 
 	result = 0;
 	negative = 1;
